+++ src/constants/constants.ts
@@ -1,0 +1,8 @@
+export const defaultColumns = [
+  'country',
+  'population',
+  'ISO',
+  'year',
+  'co2',
+  'co2_per_capita',
+];
